requirements.txt
torch
torchvision
tqdm
Pillow
matplotlib
